// src/lib/enums/leave.enums.ts
export enum LeaveType {
  ANNUAL = 'ANNUAL',
  SICK = 'SICK',
  UNPAID = 'UNPAID',
  MATERNITY = 'MATERNITY',
  PATERNITY = 'PATERNITY',
  COMPENSATORY = 'COMPENSATORY',
  BEREAVEMENT = 'BEREAVEMENT',
  STUDY = 'STUDY',
}

export enum LeaveStatus {
  PENDING = 'PENDING',
  APPROVED = 'APPROVED',
  REJECTED = 'REJECTED',
  CANCELLED_BY_USER = 'CANCELLED_BY_USER',
  CANCELLED_BY_ADMIN = 'CANCELLED_BY_ADMIN',
  TAKEN = 'TAKEN',
  PARTIALLY_TAKEN = 'PARTIALLY_TAKEN',
}

export enum HalfDayPeriod {
  FIRST_HALF = 'FIRST_HALF',
  SECOND_HALF = 'SECOND_HALF',
} 